refactor(Lotto): 객체 타입 변환

+++ src/main/java/lotto/constants/LottoRule.java
@@ -1,7 +1,7 @@
 package lotto.constants;
 
 public final class LottoRule {
-    public static final Integer NUMBER_MAX_SIZE = 6;
+    public static final Long NUMBER_MAX_SIZE = 6L;
     public static final Integer START_NUMBER = 1;
     public static final Integer END_NUMBER = 45;
